22 спринт первая реализация

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/ShoppingCartDto.java
@@ -19,5 +19,5 @@ public class ShoppingCartDto {
     UUID shoppingCartId;
 
     @NotNull
-    Map<UUID, Integer> products;
+    Map<UUID, Long> products;
 }
